Fix char, number, and flag params to use unsigned
Change function signatures to use unsigned int for values that are
never negative, eliminating unnecessary casts and type mismatches:

- Change cmd_kill(), cmd_give(), cmd_look_char() to accept unsigned int cn
  (character numbers are always positive, stored as unsigned int in map struct)
- Change cmd_some_spell() to accept unsigned int chr parameter
  (character numbers are always positive)
- Change get_near_item() flag parameter from int to unsigned int
  (flags are bitmasks, always positive values)

Also update call sites:
- Remove unnecessary casts when passing map[chrsel].cn (now matches type)
- Add explicit casts when reading character numbers from menu.opt1[n]
  (opt1 is int but stores unsigned character numbers when used for
   character commands)

This ensures the type system correctly represents that these values
cannot be negative, improving type safety and eliminating sign
conversion warnings.

+++ src/gui/context.c
@@ -380,16 +380,16 @@ int context_click(int mx, int my)
 				cmd_look_item(menu.opt1[n], menu.opt2[n]);
 				break;
 			case CMD_CHR_ATTACK:
-				cmd_kill(menu.opt1[n]);
+				cmd_kill((unsigned int)menu.opt1[n]);
 				break;
 			case CMD_CHR_GIVE:
-				cmd_give(menu.opt1[n]);
+				cmd_give((unsigned int)menu.opt1[n]);
 				break;
 			case CMD_CHR_LOOK:
-				cmd_look_char(menu.opt1[n]);
+				cmd_look_char((unsigned int)menu.opt1[n]);
 				break;
 			case CMD_CHR_CAST_K:
-				cmd_some_spell(menu.opt2[n], 0, 0, menu.opt1[n]);
+				cmd_some_spell(menu.opt2[n], 0, 0, (unsigned int)menu.opt1[n]);
 				break;
 			}
 			return 1;
